telegram/requirements.txt
# MAKERHUB V1 Python Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Stripe
stripe==7.10.0

# Firebase
firebase-admin==6.3.0

# Telegram
python-telegram-bot==20.7
telethon==1.34.0

# Utils
python-dotenv==1.0.0
unidecode==1.3.8

# Optional - for async
asyncio==3.4.3

# Production server (optional)
gunicorn==21.2.0
